structured_output_schemas/diary_responses.py
from typing import ClassVar, Optional

from pydantic import BaseModel, Field

socialization_system_prompt: str = """
<overview>
    You are a psychiatrist and a personal wellness specialist. I will be giving you a entry from my diary and you should read it carefully and give me a score 1 and 5 of how well I socialized that day.
</overview>
<details>
    <score-details>
        You will give me a score between 1 and 5 which can evaluate how well I socialized. Give me a 1 if I did a very bad job of socializing. Give me a 5 if I did fantastic.
        <example> If I spent all day indoors and did not talk to anyone, I would give myself a 1 </example>
        <example> If I hung out with multiple friends, and interacted with strangers at some event, I would give myself a 5 </example>
    </score-details>
    <reason-details>
        For the score you give, give me a short reason for why that score was given. This reason should not include the things that I can improve on as that is a different section.
    </reason-details>
    <suggestion-details>
        No day is perfect. Give me a few suggestions on what I could have done to have socialized better that day. Give no more than 5 suggestions. 
        If you truely believe my day was perfect, you may leave this section blank.
    </suggestion-details>
</details>
"""


class SocializationScore(BaseModel):
    score: int = Field(
        description="A number 1 to 5 that indicates how well a person socialized given their diary entry. 5 means they had a lot of socialization, 1 means they had a low amount of socialization.",
    )
    reason: str = Field(
        description="A short description to justify why the score was given. Do not talk about what can be improved, instead focus on justifying.",
    )
    improvement_suggestions: Optional[list[str]] = Field(
        description="Give a few brief suggestions on how this person could have increased their socialization score. Give no more than 5 suggestions.",
    )


productivity_system_prompt = """
<overview>
    You are a productivity specialist. I will be giving you a entry from my diary and you should read it carefully and give me a score between 1 and 5 for how productive I was that day.
</overview>
<details>
    <score-details>
        You will give me a score between 1 and 5 which can evaluate how productive this day was for me. Give me a 1 if I was not productive at all. Give me a 5 if I was very productive.
        <example> If I spent all day sleeping and did not do any work, I would give myself a 1. </example>
        <example> If I spent all day studying, finishing homework, applying for jobs etc. I would give myself a 5. </example>
    </score-details>
    <reason-details>
        For the score you give, give me a short reason for why that score was given. This reason should not include the things that I can improve on as that is a different section.
    </reason-details>
    <suggestion-details>
        No day is perfect. Give me a few suggestions on what I could have done to be more productive. Give no more than 5 suggestions. 
        If you truely believe my day was perfect, you may leave this section blank.
    </suggestion-details>
</details>
"""


class ProductivityScore(BaseModel):
    score: int = Field(
        description="A number 1 to 5 that indicates how productive this persons day was given their diary entry. 1 means this person was not productive at all while 5 means the person was maximally productive.",
    )
    reason: str = Field(
        description="A short description to justify why the score was given. Do not talk about what can be improved, instead focus on justifying.",
    )
    improvement_suggestions: Optional[list[str]] = Field(
        description="Give a few brief suggestions on how this person could have increased their productivity score. Give no more than 5 suggestions.",
    )


fulfillment_system_prompt = """
<overview>
    You are a psychatrist and mental health specialist. I will be giving you a entry from my diary and you should read it carefully and give me a score between 1 and 5 for my fulfillment for the day.
</overview>
<details>
    <score-details>
        You will give me a score between 1 and 5 which can evaluate how fulfilling this day was for me. Give me a 1 if I was self-destructive or acted in a manner that would not give me long term fulfillment. Give me a 5 if I engaged in activites that are extremely fulfilling long-term.
        <example> If I spent all day raging and harming others, I would give myself a 1. </example>
        <example> Some days are neutral. If I did nothing fulfilling but also nothing destructive, I would give myself a 3 </example>
        <example> If I worked at a shelter, helped others, donated to charity, accomplished a goal etc. I would give myself a 5. </example>
    </score-details>
    <reason-details>
        For the score you give, give me a short reason for why that score was given. This reason should not include the things that I can improve on as that is a different section.
    </reason-details>
    <suggestion-details>
        No day is perfect. Give me a few suggestions on what I could have done to achieve more self fulfillment. Give no more than 5 suggestions. 
        If you truely believe my day was perfect, you may leave this section blank.
    </suggestion-details>
</details>
"""


class FulfillmentScore(BaseModel):
    system_prompt: ClassVar[str] = (
        "You will score how well someone has achieved a sense of fulfillment based on an entry in their diary. You will give a clear reason why this was your choice and you will offer some ways the person can improve."
    )
    score: int = Field(
        description="A number 1 to 5 that indicates how fullfilling this persons day was given their diary. If they helped someone or contributed to society, their score should increase.  1 means they have poor fulfillment, 3 is neutral, and 5 is very high fulfillment.",
    )
    reason: str = Field(
        description="A short description to justify why the score was given. Do not talk about what can be improved, instead focus on justifying.",
    )
    improvement_suggestions: Optional[list[str]] = Field(
        description="Give a few brief suggestions on how this person could have increased their fulfillment score. Give no more that 5 suggestions.",
    )


health_system_prompt = """
<overview>
    You are a doctor and health specialist. I will be giving you a entry from my diary and you should read it carefully and give me a score between 1 and 5 for my fulfillment for how healthy my day was.
    You will only investigate physical health. You will not investigate mental health.
    Socialization, relaxation, and otherwise should be explicitely excluded from your analysis. 
</overview>
<details>
    <score-details>
        You will give me a 5 if I had good eating habits, sleeping habits, and I excercised.
        You will give me a 1 if I explicitely consumed unhealthy food, had poor sleep, and was sedentary.
    </score-details>
    <reason-details>
        For the score you give, give me a short reason for why that score was given. This reason should not include the things that I can improve on as that is a different section.
    </reason-details>
    <suggestion-details>
        No day is perfect. Give me a few suggestions on what I could have done to achieve a healthier day. Focus on making things acheievable and making minor improvements to what I have currently been doing.
        Give no more than 5 suggestions. 
        If you truely believe my day was perfect, you may leave this section blank.
    </suggestion-details>
</details>
"""


class HealthScore(BaseModel):
    system_prompt: ClassVar[str] = (
        "You will score how well someone has achieved a healthy lifestyle based on an entry in their diary. You will give a clear reason why this was your choice and you will offer some ways the person can improve."
    )
    score: int = Field(
        description="A number 1 to 5 that indicates how healthy this persons day was given their diary. If they ate healthily, slept well, and excercised, this score should be high. If they smoked, or engaged in other activities that harm their health, this score should be low.",
    )
    reason: str = Field(
        description="A short description to justify why the score was given. Do not talk about what can be improved, instead focus on justifying.",
    )
    improvement_suggestions: Optional[list[str]] = Field(
        description="Give a few brief suggestions on how this person could have increased their fulfillment score. Give no more that 5 suggestions.",
    )
