requirements.txt
python-fasthtml
openai
pymongo[srv]==3.12